Getting rid of blocks
Blocks are brittle and tend to break every now and then.
It's also extremely hard to anticipate all the edge cases
they produce. The only thing they bring is e.g. different
highlighting of the `where` keyword. The commit unifies
the highlighting of `where` and `let` since they serve
the same purpose.