Add option g:haskell_indent_let_no_in
This addresses some of the grievances in #85. Namely, we give users the
option to change how indentation works with `let` statements.

Current behavior:

    -- g:haskell_indent_let
    let x = 0
    >>>>y = 0

    let x = 0 in
    >>>>x

    -- g:haskell_indent_in
    let x = 0
    >in x

    -- ??? (conflated with g:haskell_indent_let)
    let x = 0
    >>>>x

New behavior:

    -- g:haskell_indent_no_in
    let x = 0
    >>>>x

So for example, those who omit the 'in' on the first line and any 'let'
on the second line can opt to have no indentaton:

    --- g:haskell_indent_no_in = 0
    let x = 0
    fooBar x